chore: remove workspace
