+++ .github/workflows/main.yaml
@@ -70,7 +70,7 @@ jobs:
         uses: actions/checkout@v2
 
       - name: Generate code coverage
-        run: cargo +nightly tarpaulin --verbose --engine llvm --no-dead-code --all-features --workspace --timeout 120 --out xml
+        run: cargo +nightly tarpaulin --all-features --workspace --timeout 120 --out xml
 
       - name: Upload to codecov.io
         uses: codecov/codecov-action@v2
