+++ .github/workflows/pypublish.yml
@@ -32,13 +32,13 @@ jobs:
           target: ${{ matrix.target }}
           args: --sdist --release --out dist --find-interpreter
           sccache: 'true'
-          working-directory: vcf2parquet-py
+          working-directory: python
           manylinux: auto
       - name: Upload wheels
         uses: actions/upload-artifact@v4
         with:
           name: wheels
-          path: vcf2parquet-py/dist
+          path: python/dist
 
   windows:
     runs-on: windows-latest
@@ -57,12 +57,12 @@ jobs:
           target: ${{ matrix.target }}
           args: --sdist --release --out dist --find-interpreter
           sccache: 'true'
-          working-directory: vcf2parquet-py
+          working-directory: python
       - name: Upload wheels
         uses: actions/upload-artifact@v4
         with:
           name: wheels
-          path: vcf2parquet-py/dist
+          path: python/dist
 
   macos:
     runs-on: macos-latest
@@ -80,12 +80,12 @@ jobs:
           target: ${{ matrix.target }}
           args: --sdist --release --out dist --find-interpreter
           sccache: 'true'
-          working-directory: vcf2parquet-py
+          working-directory: python
       - name: Upload wheels
         uses: actions/upload-artifact@v4
         with:
           name: wheels
-          path: vcf2parquet-py/dist
+          path: python/dist
 
   release:
     name: Release
