+++ Cargo.toml
@@ -13,32 +13,30 @@ keywords = ["bioinformatics", "parquet"]
 
 
 [dependencies]
-vcf2parquet-lib = { path = "vcf2parquet-lib", version = "0.6.0", optional = true }
-vcf2parquet-bin = { path = "vcf2parquet-bin", version = "0.6.0", optional = true }
+# parallel
+rayon      = { version = "1" }
+
+# input output management
+niffler     = { version = "2" }
+noodles     = { version = "0.64", features = ["vcf"] }
+arrow2      = { version = "0.18", features = ["io_parquet", "io_parquet_compression"] }
+rustc-hash  = { version = "1" }
+
+# logging management
+log         = { version = "0.4" }
+
+# error management
+thiserror       = { version = "1" }
+
+# cli management
+clap            = { version = "4", features = ["derive"] }
+
 
 [dev-dependencies]
+lazy_static     = { version = "1" }
 tempfile        = { version = "3" }
 assert_cmd      = { version = "2" }
 
-
-[workspace]
-members = ["vcf2parquet-lib", "vcf2parquet-bin", "vcf2parquet-py"]
-
-
-[features]
-default = ["lib"]
-
-lib = ["vcf2parquet-lib"]
-bin = ["vcf2parquet-bin"]
-
-
-[[bin]]
-name = "vcf2parquet"
-required-features = ["bin"]
-
-[[test]]
-name = "functional"
-required-features = ["bin"]
 
 [package.metadata.docs.rs]
 all-features = true
