--- vcf2parquet-bin/Cargo.toml
@@ -1,18 +1,0 @@
-[package]
-name = "vcf2parquet-bin"
-version = "0.6.0"
-edition = "2021"
-
-[dependencies]
-vcf2parquet-lib = { path = "../vcf2parquet-lib", version = "0.6.0" }
-niffler         = { version = "2" }
-arrow2          = { version = "0.18", features = ["io_parquet", "io_parquet_compression"] }
-
-# logging management
-simplelog       = { version = "0.12" }
-
-# error management
-thiserror       = { version = "1" }
-
-# cli management
-clap            = { version = "4", features = ["derive"] }
