--- vcf2parquet-lib/Cargo.toml
@@ -1,21 +1,0 @@
-[package]
-name = "vcf2parquet-lib"
-version = "0.6.0"
-edition = "2021"
-
-[dependencies]
-rayon      = { version = "1" }
-
-# input output management
-niffler     = { version = "2" }
-noodles     = { version = "0.64", features = ["vcf"] }
-arrow2      = { version = "0.18", features = ["io_parquet", "io_parquet_compression"] }
-rustc-hash  = { version = "1" }
-
-# logging and error management
-log         = { version = "0.4" }
-thiserror   = { version = "1" }
-
-[dev-dependencies]
-lazy_static = { version = "1" }
-tempfile    = { version = "3" }
